requirements.txt

Django==4.1.7
channels==4.0.0 # channels['daphne']
channels_redis==4.0.0
djangorestframework
drf-nested-routers
djoser
pycryptodome

drf-yasg # for documentation
django-debug-toolbar # for development
daphne # channels['daphne']
